improve player rendering
